Add v8 compatibility for NodeJS v12.14.1
:Release Notes:
Add v8 compatibility for NodeJS v12.14.1

:Detailed Notes:
This commit modifies the code to adapt v8 API usage based on
NodeJS v12.14.1. Compile error is fixed with modification, but
still crash happenes during dlopen this addon. Since it was
decided not to use pmloglib addon, but depends on nodejs console
log, mock addon written by js is created to keep the compatibility
where pmloglib is still required.

:Testing Performed:
Build is successfully done.
Requiring this addon introduces crash during dlopen
Verification of this addon is not performed since it was decided
not to use this module.

:QA Notes:

:Issues Addressed:
[PLAT-114757] Fix compile error of nodejs-module-webos-pmlog
              based on nodejs v12.14.1

Change-Id: Ic36fab016153c1fa49451a77628d4bc495b44714